Moved toggler position
